% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PakPC.R
\name{PakPC}
\alias{PakPC}
\title{PakPC}
\usage{
PakPC()
}
\value{
Pivot tables and graphs of PakPC2017 and PakPC2023 R packages data
}
\description{
Shiny App to Analyze Pakistan's Population Census Data
}
\examples{
if(interactive()) {
    library(PakPC)
    PakPC()
  }

}
